refactor: move maiTools api to class
